Revert "Merge 2.3.3 dev to business-dev (#292)" (#337)
This reverts commit e2ef3ad65dd33fbac655977dac3f4097c58c8015.

+++ seatunnel-connectors-v2/connector-cdc/connector-cdc-mysql/src/main/java/org/apache/seatunnel/connectors/seatunnel/cdc/mysql/config/MySqlSourceConfigFactory.java
@@ -89,7 +89,7 @@ public class MySqlSourceConfigFactory extends JdbcSourceConfigFactory {
 
         if (serverIdRange != null) {
             props.setProperty("database.server.id.range", String.valueOf(serverIdRange));
-            long serverId = serverIdRange.getServerId(subtaskId);
+            int serverId = serverIdRange.getServerId(subtaskId);
             props.setProperty("database.server.id", String.valueOf(serverId));
         }
         if (databaseList != null) {
